Alarm_Receiver/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
PyYAML>=6.0.1,<7
prometheus-client>=0.21,<1
pydantic>=2,<3
python-multipart>=0.0.9
